+++ swimmer_forward_data.py
@@ -31,8 +31,8 @@ def generate_sample(env, model, mode):
         if done:
             break
     env.close()
-    x = torch.tensor(x, dtype=torch.float32).unsqueeze(-1)
-    y = torch.tensor(y, dtype=torch.float32)
+    x = torch.tensor(x, dtype=torch.float64).unsqueeze(-1)
+    y = torch.tensor(y, dtype=torch.float64)
     return x, y
